Add same Product other vendor Search Feature to Retrieve API Endpoint
This change adds a feature to the retrieve API endpoint that allows users to search for the same product from other vendors. The code retrieves the product instance, then performs a search on all other vendors' products to find same ones. This search is based on the product's title and uses the Django's TrigramSimilarity and SearchRank to sort the search results. The code returns the original product's details along with a list of similar products from other vendors. This new feature provides users with a way to discover similar products and helps to improve the user experience of the website.