Add "product_post_save" signal and "product_retrieved" signal
This commit adds two signals to the application. The product_post_save signal updates the desired_price_reached attribute for related Favorite instances and sends an email notification to users when the price for a product has reached their desired price. The product_retrieved signal is used to increase the view count for a product whenever it is retrieved by a user. This helps to track the popularity of a product and provide insight into which products are most viewed by users.